[TensorExpr] Handle not-specified dtypes and strides.
Currently JIT has a pass to propagate shapes, but doesn't have a
capability to fill in strides and dtypes. This PR works around that by
assuming default dtype to be Float and strides corresponding to
contiguous layout, unless otherwise specified. Ideally, we won't need
this, and this is done simply as a workaround unless the corresponding
features are implemented on JIT side.

This is required for AOT compilation of mobilenet v3 with NNC.

ghstack-source-id: 0dd3d1b16e650cebb356421429fd58d38f0f003b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59346