Stop warning on .names() access in max_pool2d and max_pool2d_backward
Fixes #60053.

The problem is that `.names()` always triggers the named tensor warning.
To not trigger it, one has to guard it with has_names:
`x.has_names() ? x.names() : DimnameList{}`

This is not the first time this has happened; we should probably
make it so that .names() doesn't raise a warning unless it is actually
populated with names. That's a little tricky to implement so I'm leaving
it for the future.

Test Plan:
- New test, also run `python test/test_nn.py -v -k "max_pool"` and
confirm there are no warnings.

ghstack-source-id: fa62842f7a6ad11f4b21da060962c3d6dc015250
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60059